pyproject.toml
[project]
name = "django-df-odoo"
version = "0.0.1"
description = "Django DF ODOO"
readme = "README.md"
authors = [{name = "Apexive OSS", email = "open-source@apexive.com"}]
license = { file = "LICENSE" }
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Topic :: Internet"
]
requires-python = ">=3.9"
urls = { homepage = "https://apexive.com/" }

dependencies = [
    "Django>=4.2.0",
    "djangorestframework>=3.12.0",
    "django-df-api-drf>=1.0.0",
    "odoorpc",
    "Pillow",
    "django-environ",
    "django-hashid-field",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-django",
    "django-stubs[compatible-mypy]",
    "black>=23.7.0",
    "ruff>=0.0.284",
    "httpretty",
]

[build-system]
requires = ['setuptools>=68.1.0', 'wheel']
build-backend = 'setuptools.build_meta'

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include =  ["df_odoo*"]

[tools.black]
max-line-length = 88


[tool.ruff]
line-length = 79
fix = true


# Enable Flake's "E" and "F" codes by default.
select = ["E", "F", "I", "W", "A", "B", "Q", "C", "S"]

ignore = ["B008", "E501", "S101", "A003", "B904"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    'static',
    'migrations',
    '__pycache__',
    '.pytest_cache',
    '__init__.py',
]


# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.9.x
target-version = "py39"


[tool.ruff.per-file-ignores]
"tests/*" = ["S105", "S106"]

[tool.mypy]

ignore_missing_imports = true
disable_error_code = "attr-defined, valid-type"
disallow_untyped_defs = true
mypy_path = "df_odoo"
exclude = "venv|migrations|build|dist|docs"

plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.pytest.ini_options]

python_files = "tests.py test_*.py"
DJANGO_SETTINGS_MODULE = "tests.settings"
